moving general models to new shared library
