+++ libs/data-access-redux/src/actions/movie.actions.ts
@@ -1,7 +1,7 @@
 import { createAction } from 'redux-api-middleware';
 
 import { IMovie } from '../models/movie.model';
-import { MovieFilter } from '../models/MovieFilter';
+import { MovieFilter } from '../../../shared/models/src/MovieFilter';
 
 const BASE_URL = `${process.env.NX_API_BASE_URL}movie`;
