+++ libs/data-access-redux/src/models/LanguageModel.ts
@@ -1,4 +1,4 @@
-import { NameEntityModel } from './NameEntityModel';
+import { NameEntityModel } from '@giron/shared-models';
 
 export interface LanguageModel extends NameEntityModel {
   nameSwedish: string;
