+++ libs/data-access-redux/src/models/index.ts
@@ -4,12 +4,8 @@ export type { LanguageModel } from './LanguageModel';
 export type { IMovie, MovieFormatInfo, MoviePersonalInfo } from './movie.model';
 export { movieToString } from './movie.model';
 export type { MovieGenreModel } from './MovieGenreModel';
-export type { NameEntityModel } from './NameEntityModel';
 export type { Pageable, Sort } from './Pageable';
 export type { PersonRoleModel } from './PersonRoleModel';
 export { RoleEnum } from './RoleEnum';
-export type { SelectableModel } from './SelectableModel';
 export type { StoredData } from './storage.model';
 export { STORAGE_TYPE } from './storage.model';
-export {Filter} from './Filter';
-export {MovieFilter} from './MovieFilter';
