+++ libs/data-access-redux/src/models/movie.model.ts
@@ -2,8 +2,8 @@ import { CastAndCrewModel } from './CastAndCrewModel';
 import { CoverModel } from './CoverModel';
 import { LanguageModel } from './LanguageModel';
 import { MovieGenreModel } from './MovieGenreModel';
-import { NameEntityModel } from './NameEntityModel';
-import { SelectableModel } from './SelectableModel';
+import { NameEntityModel } from '../../../shared/models/src/NameEntityModel';
+import { SelectableModel } from '@giron/shared-models';
 
 export interface IMovie {
   id?: number;
