+++ libs/feature-movie-list-redux/src/lib/components/movie-filter/ExtendedFilterContent.tsx
@@ -3,11 +3,10 @@ import { ChangeEvent, ReactNode, useEffect, useState } from 'react';
 import {
   BaseDataStateModel,
   loadFormats,
-  MovieFilter as MovieModelFilter,
   MovieListStateModel,
-  SelectableModel,
 } from '@giron/data-access-redux';
 import { connect } from 'react-redux';
+import { MovieFilter as MovieModelFilter, SelectableModel } from '@giron/shared-models';
 
 interface ExtendedFilterContentProps {
   filter: MovieModelFilter;
