+++ libs/feature-movie-list-redux/src/lib/components/movie-filter/RegularFilterContent.tsx
@@ -1,14 +1,8 @@
 import { ChangeEvent, KeyboardEvent, ReactNode, useEffect, useState, } from 'react';
 import { FilterItem, FilterItemText, Loader } from '@giron/shared-ui-library';
 import { connect } from 'react-redux';
-import {
-  BaseDataStateModel,
-  loadGenres,
-  MovieFilter as MovieModelFilter,
-  MovieListStateModel,
-  SelectableModel,
-  updateFilter,
-} from '@giron/data-access-redux';
+import { BaseDataStateModel, loadGenres, MovieListStateModel, updateFilter, } from '@giron/data-access-redux';
+import { MovieFilter as MovieModelFilter, SelectableModel } from '@giron/shared-models';
 
 const helpFilterFreetext =
   'Du kan välja att filtrera på filmens titel, genre eller annat. Tryck sedan på ' +
