+++ libs/feature-movie-redux/src/lib/components/panels/FormatPanel.tsx
@@ -13,9 +13,9 @@ import {
   loadFormats,
   loadLanguages,
   MovieStateModel,
-  SelectableModel,
   updateMovieState,
 } from '@giron/data-access-redux';
+import { SelectableModel } from '@giron/shared-models';
 
 const REGIONS: Array<number> = [1, 2, 3, 4, 5, 6];
 const regionOptions: ReactNode[] = REGIONS.map((r: number, i: number) => {
