+++ libs/shared/models/README.md
@@ -1,0 +1,7 @@
+# shared-models
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-models` to execute the unit tests via [Jest](https://jestjs.io).
