+++ libs/shared/models/src/NameEntityModel.ts
