+++ libs/shared/models/src/SelectableModel.ts
