+++ libs/shared/models/src/index.ts
@@ -1,0 +1,4 @@
+export {Filter} from './Filter';
+export {MovieFilter} from './MovieFilter';
+export type {NameEntityModel} from './NameEntityModel';
+export type {SelectableModel} from './SelectableModel';
